+++ CHANGELOG.md
@@ -3,7 +3,44 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-# [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v13.4.1...HEAD)
+# [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v14.0.0...HEAD)
+
+# [v14.0.0](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v13.4.1...v14.0.0)
+### Breaking Changes
+* HttpClient response objects now use `data` instead of `body`. For example, if you make a request like this:
+```ts
+const response = await HttpClient.requestApi('/users/', { method: 'get' })
+
+// old way which is no longer valid:
+const users = response.body;
+
+// new way:
+const users = response.data;
+```
+### Added
+* [PCM-1837](https://inindca.atlassian.net/browse/PCM-1837) – add `setAccessToken(token)` function
+* [PCM-1844](https://inindca.atlassian.net/browse/PCM-1844) – stop sending logs on disconnect:
+  * add `stopServerLogging()` & `startServerLogging()` functions to allow consumers to stop sending server
+    logs that are sent via the client-logger.
+  * on `client.disconnect()`, logs will stop being sent to the server.
+  * on `client.connect()`, logs will start being sent to the server again.
+
+### Changed
+* [PCM-1842](https://inindca.atlassian.net/browse/PCM-1842) – migrate to the new pipeline. Also versioning cdn urls with major and exact versions. For example:
+    * `/v13.5.0/streaming-client.browser.js` (exact version)
+    * `/v13/streaming-client.browser.js` (locked to latest for a specific major version)
+* [PCM-1842](https://inindca.atlassian.net/browse/PCM-1842)/[PCM-1560](https://inindca.atlassian.net/browse/PCM-1560) – Upgrade to new pipeline
+
+### Fixed
+* [ACE-2053](https://inindca.atlassian.net/browse/ACE-2053) – Remove superagent which is no longer maintained in order to get away from the 'formidable' snyke vulnerability.
+* [PCM-1908](https://inindca.atlassian.net/browse/PCM-1908) – fixing some `.connect()` functionality:
+    * `autoReconnect` no longer default to `true` but will be set to true after successfully connecting once
+    * when `connect()` times out, it will call through to stop any pending WS connect that stanza my still be attempting
+    * `connect()` will now reject when stanza emits a `--transport-disconnected` event which is what stanza emits when there
+        was a WS connection that failed or terminated. Note that stanza does not sufface the error, so we will be rejecting
+        with a generic error.
+* Addressed snyk and npm audit issues
+* [PCM-1862](https://inindca.atlassian.net/browse/PCM-1862) - remove individual topics from the tracked lists (subscriptions) after their last handlers have been removed. Fixed `_notifications.resubscribe()` to not treat individual topics as bulk topics
 
 # [v13.4.1](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v13.4.0...v13.4.1)
 ### Added